requirements.txt
Flask
pdf2image
PyPDF2
docx2pdf
Werkzeug
python-docx
Pillow